rustc_metadata: Use a query for collecting all traits in encoder
